feat: move helpers to utils
